Simple Command Line Calculator
Simple Command Line Calculator by Zohaib Javed is a streamlined, easy-to-use calculator application designed for the command line interface. Leveraging the robustness of Node.js and the strong typing provided by TypeScript, this project introduces a user-friendly experience with the help of Inquirer for managing user input. Whether you need to perform basic arithmetic operations or more complex calculations, this tool is crafted to provide precision and ease. Adhering to open-source principles, it is available under the MIT License, ensuring that it can be freely used, modified, and distributed.